@Valid 사용해서 예외처리 사용하지않는 검증 함수 삭제

+++ src/main/java/org/example/schedulemanagement/repository/CommentRepository.java
@@ -10,5 +10,5 @@ import java.util.List;
 public interface CommentRepository extends JpaRepository<Comment, Long> {
     void deleteByScheduleId(Long scheduleId);
     List<Comment> findByScheduleId(Long scheduleId);
-    Long countByScheduleId(Long scheduleId);
+    int countByScheduleId(Long scheduleId);
 }
